requirements.txt
asgiref==3.2.10
Django==3.0.8
django-markdownx==3.0.1
django-taggit==1.3.0
html5lib==1.1
Markdown==3.2.2
Pillow==7.2.0
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.44
six==1.15.0
sqlparse==0.3.1
Wand==0.6.2
webencodings==0.5.1
xhtml2pdf==0.2.4
